fix cta buttons, selects
