+++ apps/dotcom/client/src/tla/components/TlaEditor/TlaEditorTopRightPanel.tsx
@@ -1,11 +1,12 @@
+import { SignInButton } from '@clerk/clerk-react'
 import classNames from 'classnames'
 import { useRef } from 'react'
 import { PeopleMenu, usePassThroughWheelEvents } from 'tldraw'
 import { useCurrentFileId } from '../../hooks/useCurrentFileId'
+import { F } from '../../utils/i18n'
+import { TlaCtaButton } from '../TlaCtaButton/TlaCtaButton'
 import { TlaFileShareMenu } from '../TlaFileShareMenu/TlaFileShareMenu'
-import { TlaIcon } from '../TlaIcon/TlaIcon'
-import { TlaShareButton } from '../TlaShareButton/TlaShareButton'
-import { TlaSignUpButton } from '../TlaSignUpButton/TlaSignUpButton'
+import { TlaSignedOutShareButton } from '../TlaSignedOutShareButton/TlaSignedOutShareButton'
 import styles from './top.module.css'
 
 export function TlaEditorTopRightPanel({
@@ -23,14 +24,16 @@ export function TlaEditorTopRightPanel({
 		return (
 			<div ref={ref} className={classNames(styles.topRightPanel)}>
 				<PeopleMenu displayUserWhenAlone={false} />
-				<div className={styles.signInButtons}>
-					<TlaFileShareMenu fileId={fileId!} context={context} source="anon">
-						<button data-testid="share-button" className={classNames(styles.shareButtonMini)}>
-							<TlaIcon icon="share" />
-						</button>
-					</TlaFileShareMenu>
-					<TlaSignUpButton />
-				</div>
+				<TlaSignedOutShareButton fileId={fileId} context={context} />
+				<SignInButton
+					mode="modal"
+					forceRedirectUrl={location.pathname + location.search}
+					signUpForceRedirectUrl={location.pathname + location.search}
+				>
+					<TlaCtaButton>
+						<F defaultMessage="Sign in" />
+					</TlaCtaButton>
+				</SignInButton>
 			</div>
 		)
 	}
@@ -39,7 +42,9 @@ export function TlaEditorTopRightPanel({
 		<div ref={ref} className={styles.topRightPanel}>
 			<PeopleMenu displayUserWhenAlone={false} />
 			<TlaFileShareMenu fileId={fileId!} source="file-header" context={context}>
-				<TlaShareButton />
+				<TlaCtaButton>
+					<F defaultMessage="Share" />
+				</TlaCtaButton>
 			</TlaFileShareMenu>
 		</div>
 	)
