+++ apps/dotcom/client/src/tla/components/TlaFileError/TlaFileError.tsx
@@ -1,9 +1,10 @@
+import { SignInButton } from '@clerk/clerk-react'
 import { TLRemoteSyncError, TLSyncErrorCloseEventReason } from '@tldraw/sync-core'
 import { ReactElement, useEffect } from 'react'
 import { sadFaceIcon } from '../../../components/ErrorPage/ErrorPage'
 import { useSetIsReady } from '../../hooks/useIsReady'
 import { F } from '../../utils/i18n'
-import { TlaSignUpButton } from '../TlaSignUpButton/TlaSignUpButton'
+import { TlaCtaButton } from '../TlaCtaButton/TlaCtaButton'
 import styles from './TlaFileError.module.css'
 
 function DefaultError() {
@@ -49,7 +50,17 @@ export function TlaFileError({ error }: { error: unknown }) {
 				<TlaFileErrorContent
 					header={<F defaultMessage="Private file" />}
 					para1={<F defaultMessage="Contact the owner to request access." />}
-					cta={<TlaSignUpButton />}
+					cta={
+						<SignInButton
+							mode="modal"
+							forceRedirectUrl={location.pathname + location.search}
+							signUpForceRedirectUrl={location.pathname + location.search}
+						>
+							<TlaCtaButton>
+								<F defaultMessage="Sign in" />
+							</TlaCtaButton>
+						</SignInButton>
+					}
 				/>
 			)
 		}
