+++ apps/dotcom/client/src/tla/components/TlaSelect/select.module.css
@@ -4,7 +4,9 @@
 	display: flex;
 	align-items: center;
 	height: 100%;
+	gap: 4px;
 	justify-content: flex-end;
+	cursor: pointer;
 }
 
 .wrapper {
@@ -27,7 +29,9 @@ div:has(> .content) {
 .option {
 	display: flex;
 	align-items: center;
-	padding: 8px;
+	padding: 8px 8px 8px 12px;
+	gap: 4px;
+	cursor: pointer;
 }
 
 @media (hover: hover) {
