+++ apps/dotcom/client/src/tla/components/TlaSignedOutShareButton/signed-out-share-button.module.css
@@ -1,0 +1,34 @@
+.signedOutShareButton {
+	height: 40px;
+	width: 40px;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+	background: none;
+	border: none;
+	position: relative;
+	cursor: pointer;
+	z-index: 1;
+	color: var(--color-text-1);
+}
+
+@media (hover: hover) {
+	.signedOutShareButton::after {
+		display: block;
+		content: '';
+		position: absolute;
+		top: 4px;
+		left: 4px;
+		right: 4px;
+		bottom: 4px;
+		border-radius: 4px;
+		background-color: var(--tla-color-hover-1);
+		z-index: 0;
+		pointer-events: none;
+		visibility: hidden;
+	}
+
+	.signedOutShareButton:hover::after {
+		visibility: visible;
+	}
+}
